allow user to override 'go get' for private repos
this allows us to put private repos that are inaccessible to 'go get' in the
Gomfile. an example:

  $ cat Gomfile
  [...]
  gom 'github.com/ActiveState/logyard', :command => 'git clone http://git-mirrors.activestate.com/logyard.git vendor/src/github.com/ActiveState/logyard', :branch => 'cfv2'
  [...]

we maintain a mirror of our github private repos, and we want Gom to clone from this mirror url (gom will also be invoked from packaging systems, so reliability is important).

i'd like to see if this syntax can be improved. for example, i don't like having to specify the local clone path "vendor/src/github.com/...".